fix: more uniform types

+++ packages/smartcontracts/contracts/BrokerEconomics/policies/IAllocationPolicy.sol
@@ -3,9 +3,9 @@
 pragma solidity ^0.8.13;
 
 interface IAllocationPolicy {
-    function setParam(uint256 param) external;
+    function setParam(uint param) external;
     function calculateAllocation(address broker) external returns (uint allocation);
-    function getHorizonSeconds() external view returns (uint256 horizonSeconds);
+    function getHorizonSeconds() external view returns (uint horizonSeconds);
     function onJoin(address broker) external;
     function onLeave(address broker) external;
     function onStakeIncrease(address broker) external;
